+++ koin-projects/koin-core/src/main/kotlin/org/koin/core/definition/Options.kt
@@ -20,4 +20,4 @@ package org.koin.core.definition
  *
  * @author Arnaud Giuliani
  */
-data class Options(var isCreatedAtStart: Boolean = false, var override: Boolean = false)
+data class Options(var isCreatedAtStart: Boolean = false, var override: Boolean = false, var isExtraDefinition : Boolean = false)