+++ koin-projects/koin-core/src/test/java/org/koin/dsl/AdditionalTypeBindingTest.kt
@@ -134,29 +134,26 @@ class AdditionalTypeBindingTest {
         koin.get<Simple.ComponentInterface1>(named("default"))
     }
 
-    @Test
-    fun `can resolve an additional types`() {
-        val app = koinApplication {
-            printLogger()
-            modules(
-                    module {
-                        single { Simple.Component1() } binds arrayOf(
-                                Simple.ComponentInterface1::class,
-                                Simple.ComponentInterface2::class
-                        )
-                    })
-        }
-
-        app.assertDefinitionsCount(1)
-
-        val koin = app.koin
-        val c1 = koin.get<Simple.Component1>()
-        val ci1 = koin.bind<Simple.ComponentInterface1, Simple.Component1>()
-        val ci2 = koin.bind<Simple.ComponentInterface2, Simple.Component1>()
-
-        assertEquals(c1, ci1)
-        assertEquals(c1, ci2)
-    }
+//    @Test
+//    fun `can resolve an additional types`() {
+//        val koin = koinApplication {
+//            printLogger()
+//            modules(
+//                    module {
+//                        single { Simple.Component1() } binds arrayOf(
+//                                Simple.ComponentInterface1::class,
+//                                Simple.ComponentInterface2::class
+//                        )
+//                    })
+//        }.koin
+//
+//        val c1 = koin.get<Simple.Component1>()
+//        val ci1 = koin.bind<Simple.ComponentInterface1, Simple.Component1>()
+//        val ci2 = koin.bind<Simple.ComponentInterface2, Simple.Component1>()
+//
+//        assertEquals(c1, ci1)
+//        assertEquals(c1, ci2)
+//    }
 
     @Test
     fun `additional type conflict`() {
@@ -174,7 +171,22 @@ class AdditionalTypeBindingTest {
     }
 
     @Test
-    fun `conflicting with additional types`() {
+    fun `resolve all`() {
+        val koin = koinApplication {
+            printLogger(Level.DEBUG)
+            modules(
+                    module {
+                        single<Simple.ComponentInterface1> { Simple.Component1() }
+                        single { Simple.Component2() } bind Simple.ComponentInterface1::class
+                        single { getAll<Simple.ComponentInterface1>() }
+                    })
+        }.koin
+
+        assert(koin.get<List<Simple.ComponentInterface1>>().size == 2)
+    }
+
+    @Test
+    fun `additional types`() {
         val koin = koinApplication {
             printLogger(Level.DEBUG)
             modules(
@@ -187,5 +199,19 @@ class AdditionalTypeBindingTest {
                     })
         }.koin
         assert(koin.getAll<Simple.ComponentInterface1>().size == 2)
+    }
+
+    @Test
+    fun `getAll 1 types`() {
+        val koin = koinApplication {
+            printLogger(Level.DEBUG)
+            modules(
+                    module {
+                        single<Simple.ComponentInterface1> { Simple.Component2() }
+                        single { getAll<Simple.ComponentInterface1>() }
+                    })
+        }.koin
+        assert(koin.getAll<Simple.ComponentInterface1>().size == 1)
+        assert(koin.get<List<Simple.ComponentInterface1>>().size == 1)
     }
 }